refactor(snippetz): remove httpsnippet-lite/**/rust/*, consolidate rust async/sync into single reqwest plugin
