--- packages/snippetz/src/httpsnippet-lite/types/targets/rust/reqwest/client.d.ts
@@ -1,12 +1,0 @@
-/**
- * @description
- * HTTP code snippet generator for Rust using Reqwest.
- *
- * @author
- * @brianheineman
- *
- * for any questions or issues regarding the generated code snippet, please open an issue mentioning the author.
- */
-import type { Client } from '../../targets.js'
-
-export declare const reqwest: Client
