--- packages/snippetz/src/httpsnippet-lite/types/targets/rust/target.d.ts
@@ -1,3 +1,0 @@
-import type { Target } from '../targets.js'
-
-export declare const rust: Target
