+++ packages/snippetz/src/plugins/rust/reqwest/index.ts
@@ -1,2 +1,1 @@
-export * from './async'
-export * from './sync'
+export * from './reqwest'
